# Converge IMU faster at higher groundspeed during GPS Rescue