--- a/src/main/flight/gps_rescue.c
+++ b/src/main/flight/gps_rescue.c
@@ -112,8 +112,8 @@ typedef struct {
     float alitutudeStepCm;
     float maxPitchStep;
     float absErrorAngle;
-    float groundspeedErrorRatio;
-    float dcm_kpModifier;
+    float groundspeedPitchAttenuator;
+    float imuYawGain;
 } rescueSensorData_s;
 
 typedef struct {
@@ -236,7 +236,7 @@ static void rescueAttainPosition(void)
         previousAltitudeError = 0.0f;
         throttleAdjustment = 0;
         rescueState.intent.disarmThreshold = gpsRescueConfig()->disarmThreshold * 0.1f;
-        rescueState.sensor.dcm_kpModifier = 1.0f;
+        rescueState.sensor.imuYawGain = 1.0f;
         return;
     case RESCUE_DO_NOTHING:
         // 20s of slow descent for switch induced sanity failures to allow time to recover
@@ -572,14 +572,15 @@ static void sensorUpdate(void)
     // values to modify dcm_kp to a low value when climbing, and a higher value when there is a velocity error
     if (gpsRescueConfig()->rescueGroundspeed) {
         const float setGroundspeedInv = 1.0f / gpsRescueConfig()->rescueGroundspeed;
-        const float groundspeedErrorRatio = fabsf(rescueState.sensor.groundSpeedCmS - rescueState.sensor.velocityToHomeCmS) * setGroundspeedInv;
+        const float groundspeedPitchAttenuator = fabsf(rescueState.sensor.groundSpeedCmS - rescueState.sensor.velocityToHomeCmS) * setGroundspeedInv;
         // 0 if groundspeed = velocity to home,
         // 1 if moving sideways at target groundspeed but zero home velocity,
         // 2 if moving backwards (away from home) at target groundspeed
-        rescueState.sensor.dcm_kpModifier = constrainf(1.0f + groundspeedErrorRatio, 1.0f, 6.0f); // up to 6x dcm_kp increase
-        rescueState.sensor.groundspeedErrorRatio = fminf(rescueState.sensor.dcm_kpModifier, 3.0f); // cut pitch angle by up to one third
+        // 4 if moving backwards (away from home) at twice target groundspeed
+        rescueState.sensor.imuYawGain = constrainf(1.0f + groundspeedPitchAttenuator, 1.0f, 6.0f); // up to 6x increase in IMU Yaw gain
+        rescueState.sensor.groundspeedPitchAttenuator = fminf(rescueState.sensor.imuYawGain, 3.0f); // cut pitch angle by up to one third
         const bool climbOrRotate = (rescueState.phase == RESCUE_ATTAIN_ALT) || (rescueState.phase == RESCUE_ROTATE);
-        rescueState.sensor.dcm_kpModifier = (climbOrRotate) ? 0.0f : rescueState.sensor.dcm_kpModifier;
+        rescueState.sensor.imuYawGain = (climbOrRotate) ? 0.0f : rescueState.sensor.imuYawGain;
         // zero dcm_kp during climb or rotate, to stop IMU error accumulation arising from drift during long climbs
     }
 
@@ -666,10 +667,10 @@ void descend(void)
         // if quad drifts further than 2m away from home, should by then have rotated towards home, so pitch is allowed
         rescueState.intent.rollAngleLimitDeg = gpsRescueConfig()->maxRescueAngle * rescueState.intent.proximityToLandingArea;
         // reduce roll capability when closer to home, none within final 2m
+        rescueState.intent.pitchAngleLimitDeg = gpsRescueConfig()->maxRescueAngle / rescueState.sensor.groundspeedPitchAttenuator;
     }
 
     rescueState.intent.yawAttenuator = 1.0f; // just in case entered descend phase before yaw authority was complete
-    rescueState.intent.pitchAngleLimitDeg = gpsRescueConfig()->maxRescueAngle; // just in case we entered descend directly
 
     // configure altitude step for descent, considering interval between altitude readings
     rescueState.intent.altitudeStep = -1.0f * rescueState.sensor.altitudeDataIntervalSeconds * gpsRescueConfig()->descendRate;
@@ -777,8 +778,8 @@ void gpsRescueUpdate(void)
             rescueState.intent.yawAttenuator += rescueState.sensor.gpsRescueTaskIntervalSeconds;
         }
         if (rescueState.sensor.absErrorAngle < 30.0f) {
-            // allow pitch, limiting allowed angle if there is a significant groundspeedErrorRatio
-            rescueState.intent.pitchAngleLimitDeg = gpsRescueConfig()->maxRescueAngle / rescueState.sensor.groundspeedErrorRatio;
+            // allow pitch, limiting allowed angle if there is a significant groundspeedPitchAttenuator
+            rescueState.intent.pitchAngleLimitDeg = gpsRescueConfig()->maxRescueAngle / rescueState.sensor.groundspeedPitchAttenuator;
             rescueState.phase = RESCUE_FLY_HOME; // enter fly home phase
             rescueState.intent.secondsFailing = 0; // reset sanity timer for flight home
             rescueState.intent.proximityToLandingArea = 1.0f; // velocity iTerm activated, initialise proximity for descent phase at 1.0
@@ -814,7 +815,7 @@ void gpsRescueUpdate(void)
  
         if (newGPSData) {
             // cut back on allowed angle if there is a high groundspeed error
-            rescueState.intent.pitchAngleLimitDeg = gpsRescueConfig()->maxRescueAngle / rescueState.sensor.groundspeedErrorRatio;
+            rescueState.intent.pitchAngleLimitDeg = gpsRescueConfig()->maxRescueAngle / rescueState.sensor.groundspeedPitchAttenuator;
             if (rescueState.sensor.distanceToHomeM <= rescueState.intent.descentDistanceM) {
                 rescueState.phase = RESCUE_DESCENT;
                 rescueState.intent.secondsFailing = 0; // reset sanity timer for descent
@@ -874,9 +875,9 @@ float gpsRescueGetYawRate(void)
     return rescueYaw;
 }
 
-float gpsRescueGetDcmKpModifier(void)
+float gpsRescueGetImuYawGain(void)
 {
-    return rescueState.sensor.dcm_kpModifier;
+    return rescueState.sensor.imuYawGain;
 }
 
 float gpsRescueGetThrottle(void)