Tested, works well, but would appreciate further testing please!  

Makes the following changes:

- Yaw IMU gain increases up to 6x when there is a large drift over ground error, so that the IMU can re-orient itself more quickly the worse the error
- locks yaw IMU gain to zero during climb and rotate phases, preventing acquisition of IMU orientation errors due to drift during long climbs
- attenuates the maximum angle allowed when there is a large drift over ground error, to reduce the speed of the flyaway that can happen when the IMU orientation is very incorrect
- max rescue angle defaults to 45 degrees.  Even with a maximum IMU error, 45 degrees max angle doesn't cause the quad to speed off madly into the distance.
- fixes an issue where direct entry to a descent mode may not gain full yaw authority
- changes the sensitivity of the `FLY HOME` sanity check.  Now the sanity check will only count up towards to the 15s disarm value if the quad fails to reach 10%, not 50%, of the set return speed, when checked each second.  As before this is an up-down counter, so any time the quad attains >10% of the set return velocity to home, the counter decrements by 1, towards a minimum of zero, and when less than 10%, the counter increments by one.  If there is a net cumulative period of more than 15s with less than 10% forward speed to home, the sanity check will disarm the quad, since it is most unlikely that it will return home.  Users should expect some counting up in the early part of a rescue, especially with bad IMU disorientation, but with the above changes the quad should not disarm from this sanity check unless something really has gone badly wrong, eg stuck in a tree, or crashed, or extremely high wind.

WARNING 1: 
A GPS-only rescue system cannot handle high wind drift - do NOT rely on GPS Rescue in winds above 50-60kph unless you have a properly calibrated and fully functional, tested, Magnetometer (compass).  Higher maximum angle will help handle strong winds better, but the max throttle must be increased also, and in the face of a temporary IMU error, the quad may fly off in the wrong direction at high speed.  In practice, a 35-45 degree angle is more than enough for normal conditions, a 50-55 degree angle may be better in stronger winds.  Always test the return in the conditions of the day to ensure your quad will turn and head for home effectively.  If the wind is too high, the quad will kind of spiral around, at speed, making little or no headway.  Don't let it do this for more than 15s or it will sanity check and drop.

WARNING 2:
IMU disorientation is always possible with GPS-only systems. If a rescue is initiated and the IMU is disoriented from prior drift or high wind, the quad may initially head off in completely the wrong direction, even though the arrow points straight up.  Functionally the quad thinks it is flying home, but has temporarily got the orientation of the quad wrong.  The IMU will, or should, correct itself over the next 2-3seconds, but will fly in a large arc while doing so.  If the maxim allows angle in the rescue is set high, this arc may be flown at high speed, and cover some considerable distance, but the quad will likely handle higher wind rescues a bit more reliably.

To test: enter a GPS Rescue after a 6-8s period of tail drift (fly backwards at > 2m/s for say 6-10s).  Hold that drift long enough for the arrow to rotate 180 degrees wrong.  Initiate the rescue. The rescue will then rotate entirely the wrong direction, ie fly directly away from home.  With this PR it should gain speed away from home for less than a second, and then smoothly and steadily, over 2-3 seconds rotate towards home.  The arrow will keep pointing straight up all the time (the turning is the IMU correcting itself).  After maybe 2-4s, it should be heading home.  The quad will attain a decent speed while correcting itself.  The max speed is determined mostly by the maximum angle. 

If the debug mode `RTH` is displayed on the OSD screen, the sanity check timer is the value of the last two digits of the last debug element.

Testing in high wind and/or with prior extreme back drift would be appreciated.
Testing to figure out how much wind can be tolerated would be helpful, too.

Thanks to @Sek101 for investigating and clarifying the problem.  To summarise, the problem is that:

- if the quad drifts in a direction other than in the direction of the nose of the quad, and
- the drift is at a significant speed (anything > 2m/s), and
- the drift is sustained for at least one or two dcm_kp time constants (6-8 seconds)
- the IMU will, over time, consider that the nose of the quad is pointing in the direction of the course over ground, not the true heading of the quad, and point the arrow to the bottom of the goggles, thinking that is where home is.
- Then, when a rescue is initiated, the quad will rotate so that the arrow points upwards, and will fly in a direction that can be entirely incorrect, including a full 180 degrees wrong.  It will also fly faster than usual at first, and may make a large arc in the sky before finally settling on the true path to home.

This excellent video from Sek101 shows a takeoff, then a quick rotation to fly away from home backwards.  The home arrow slowly orients itself to become 180 degrees wrong.  When the rescue is initiated, the quad rotates according to the incorrect IMU orientation, and flies off 180 degrees in the wrong direction.  
https://www.youtube.com/watch?v=DyEYdI-Qgtw

Code attempt 1 = FAILURE!
This idea was simple: make the dcm_kp value proportional to groundspeed during a GPS Rescue, using about a 4x increase in dcm_kp at normal rescue speed, and very low dcm_kp at low/zero groundspeed.  
RESULT: spirals during initiation when IMU error existed at the start, quad flies off very fast, then enters a kind of "chasing the tail" behaviour.
Video shows, around 44s, quad in reverse, arrow going wrong, then persisting rotation.
https://www.youtube.com/watch?v=BSOgY-f1XtM

Code attempt 2 = NOT QUITE THERE YET!
Here we detect an absolute error between groundspeed and velocity to home, normalised to the intended return speed, plus one.  If the quad is flying directly home, there is no velocity error, so this `groundspeedErrorRatio` value would be 0+1 , ie  1.0.  If the quad is flying off at 90 degrees to home at a groundspeed equal to the intended return speed, the `groundspeedErrorRatio` would be 2.0.  If flying off with a full 180 degree error, at the intended return speed, this factor would be 3.0.  
Then we use this factor to boost `dcm_kp` during the rescue.
RESULT: Some improvement, but still tended to spiral and pick up a lot of speed.

Code attempt 3 = very good!
- calculate `groundspeedErrorRatio` as above, and use it to boost `dcm_kp` up to 5 times.
- use the inverse of this value to attenuate the maximum allowed pitch angle, thereby limiting absolute velocity when the IMU error would send the quad in the wrong direction.
- set `dcm_kp` forcibly to zero while in the climb or rotate phases of the rescue.  This locks in the pre-existing IMU state, which is normally pitch forward flight with a high likelihood of pointing to home.  By locking in that IMU state, we prevent drift (from wind, etc) from building up an IMU error of the kind that could cause a flyaway.
- finally, restore the max angle to the previous default of 40, for pilots who may have GPS modules that aren't working well, to avoid excessive pitch error issues in those cases.

Here is some low-res video of the IMU problem and how this PR achieves a pretty good result in the face of the worst possible IMU errors.

https://www.youtube.com/watch?v=IKbeWCsQj24

This PR should significantly reduce the chance that a working GPS Rescue setup would fail, especially if there is, or has been, signficant sideways or rear-wards drift either for a period of time before the rescue was initiated, or during the climb period.  Previous versions of GPS Rescue were very susceptible to IMU disorientation and could enter persistent non-recovering flyaways.