@jsk2084 2023-05-01
Tested and found this PR to be an excellent solution to the heading error problem that can sometimes arise when not using a compass. The following clip starts with a quad flying around home point, drifting sideways then inducing full heading error by flying backwards for some time before starting a Rescue. After climb, it quickly and neatly corrects it's course towards home! Also tested with a max angle setting of 60 degrees and groundspeed of 50kph... No issues encountered so far! https://youtu.be/6MVMsMArtVo