@ctzsnooze 2023-05-02
Added a small change to the `FLY HOME` phase sanity check to only count up when return velocity to home is <25% of set velocity to home, and (sorry for going up and down recently) set the default angle to 45 degrees.  These changes should make it less likely to trigger a sanity in strong headwinds that could slow the quad down.

If, for instance, the quad travels at 80kph at 45 degree pitch angle, a headwind of 60kph would reduce net speed to home to only 20kph, or into the range that would cause a 25% sanity check threshold to fail.  Users who want to fly long distances with strong tail winds should consider a higher GPS max angle value, and consider the possible time taken to fly home against such a wind, since throttle may be high for a long time, and the battery may get very low.

Thanks to @jsk2084 for confirming that 60 degrees is OK with the changes in this PR.  Something around 70 degrees would seem to be the practical absolute upper limit.