@ctzsnooze 2023-05-02
@ledvinap thanks for your comments, I've implemented what I can.  The problem I seek to solve arises only when there is IMU disorientation, and this happens in only a very small proportion of rescues, typically those with long climb times in high wind.  There have been a number of reports and in all cases drift during the climb eventually led me to an understanding of how things went wrong.  It has been a really tough challenge to find a solution. This solution definitely works better than anything else I could come up with.  I wish we had a better solution.  If anyone can find one, that would be great.  But for now this 'works'.