@ledvinap 2023-05-02
Maybe there should be 'IMU heading align' GPS phase after climbing to altitude.
IMU attitude shall be quite fine after climb (maybe stay in climb phase for few seconds it is too short).
Set forward pitch limit to some reasonable value and speed setpoint to value where there is good GPS heading acquisition. Wait until GPS and IMU headings match, with RTH heading PID active (to track decreasing IMU error). Increase gain here if necessary.
Then enter normal RTH phase (probably speeding up to full RTH speed).