@ledvinap 2023-05-02
Also, CoG compensation is probably very poor when error angle is 180 deg. https://github.com/betaflight/betaflight/blob/master/src/main/flight/imu.c#L222 

IIRC it is calculating vector cross product, which is small for antiparallel case. It may be special-cased for error >~ 90deg, applying constant 1 instead of sin(theta) in that case.