@ctzsnooze 2023-05-03
> Maybe there should be 'IMU heading align' GPS phase after climbing to altitude.

@ledvinap it would be very nice if we could know for sure that the IMU was good, but in cases of strong wind, or a poorly trimmed accelerometer, the climb phase course over ground could be very different from the direction to home, and the IMU will then assume that the nose is pointing in the direction of the course over ground, and become very incorrect, especially for long climbs which give it plenty of time.

That's why I lock the IMU during climbs.  As you appreciate, it would be best to lock it only for yaw, I would like that better than the current situation.

In strong wind, at the end of the climb, the quad could be blown backwards at high speed.  This is an essentially impossible case to manage once we complete the climb.  If we apply a certain amount of pitch, and if that amount is inadequate to overcome the wind, even when the nose is pointing to home, then the rescue must inevitably fail.  If we apply enough pitch to overcome the wind, we have a chance; but if the IMU is incorrect and wants to fly us away from home, then when we apply that strong pitch forward, we will go extremely fast downwind.  True, the IMU will correct, at least partially starting to turn, but the arc to return is very long and very slow, and not until the nose is effectively close to pointing to home can the pitch angle succeed in making any headway against the wind.  In practice, this is why I lost a quad of mine to a sanity check recently while testing this exact situation.  So we kind of need that angle to provide speed into the wind, but adding the same angle when the nose is not true to home will be disastrous also.  By constraining the angle, especially when the velocity away is high, then we don't fly at high speed down-wind, and we don't make such a long arc when the IMU corrects, but then we don't have the forward angle to defeat the wind.

So overall I think it is simply impossible to handle a certain amount of wind, with GPS data alone.  The IMU will always be overwhelmed at some wind velocity, and without accurate knowledge of the true attitude of the quad, eg from a Mag, such rescues are doomed.

So this PR should make it so that the quad will handle some degree of wind drift, and previous IMU disorientation in reasonable air conditions, effectively.  For windy conditions it should still work, to a point, even if the return is a bit all over the place.  And we would need to accept that in strong wind, there is no chance, without good Mag code working properly.  If we have a Mag, it's easy, point the nose to home, lean forward, no problem.