@ledvinap 2023-05-03
IMO it is worth trying yaw-only gain staging during rescue. Implementation is quite trivial, I'll look into it.
In long term, better IMU algorithm, possibly even with wind estimation, will solve most of the problems (= some form of Kalman estimator ;-) ). It shouldn't be that hard to port it from other FC (but it will probably require dual sample rate to support high gyro rates. That may complicate thing a bit)