@ledvinap 2023-05-03
IMU heading align:

1. finish altitude phase (roll/pitch setpoint zero; yaw to home; cog disabled)
   - possibly pre-estimate drift from GPS change in last phase of ascent

2. roll/pitch setpoint zero; yaw fixed; altitude fixed; cog disabled
   - wait until timeout (few s) or significant GPS change (several meters, dependent on GPS resolution)
   - estimate velocity vector during this phase (both wind and ACC misalignment is estimated here)

3.  roll=0; pitch = small forward angle; yaw fixed; altitude fixed; cog disabled
    - wait until significant GPS change or safety timeout
       - subtract velocity 2 when detecting 'significant' change; this will improve error offset/wind handling
    - estimate velocity vector

4. Get forward vector (velocity 3. - velocity 2.)
   - set IMU yaw from forward vector
   - (maybe) set yaw to home, wait until error is reasonable
   - (maybe) set RTH velocity setpoint
   - enable COG

4. Continue RTH phases