@ctzsnooze 2023-05-04
Thanks @ledvinap for your thoughts, above.  Indeed it would be awesome to look at a more comprehensive, and less empirical, solution. 

For now, if we can make it work nicely in light wind, and be quite OK up to say 30-40kph winds, that would be good.   For higher wind speeds the probability of failure is likely to increase.  Maybe we then advise users who need reliable returns above a certain wind velocity to use a Magnetometer.  That's about the best I'm hoping for, at this point.